@@ -4,31 +4,30 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.TextView
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import io.github.ja2stracciatella.ConfigurationModel
-import io.github.ja2stracciatella.R
+import io.github.ja2stracciatella.databinding.FragmentLauncherBinding
 
 /**
  * A placeholder fragment containing a simple view.
  */
 class PlaceholderFragment : Fragment() {
+    private var _binding: FragmentLauncherBinding? = null
+    private val binding get() = _binding!!
 
     private lateinit var configurationModel: ConfigurationModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        val root = inflater.inflate(R.layout.fragment_launcher, container, false)
-        val textView: TextView = root.findViewById(R.id.section_label)
-        configurationModel = ViewModelProvider(requireActivity()).get(ConfigurationModel::class.java)
-        configurationModel.vanillaGameDir.observe(viewLifecycleOwner, Observer<String> {
-            textView.text = it
-        })
-        return root
+    ): View {
+        _binding = FragmentLauncherBinding.inflate(inflater, container, false)
+        configurationModel = ViewModelProvider(requireActivity())[ConfigurationModel::class.java]
+        configurationModel.vanillaGameDir.observe(viewLifecycleOwner) {
+            binding.sectionLabel.text = it
+        }
+        return binding.root
     }
 
     companion object {
