@@ -1,38 +1,36 @@
 package io.github.ja2stracciatella.ui.main
 
-import android.content.Context
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
-import androidx.fragment.app.FragmentPagerAdapter
+import androidx.fragment.app.FragmentActivity
+import androidx.viewpager2.adapter.FragmentStateAdapter
 import io.github.ja2stracciatella.R
 
-private val TAB_TITLES = arrayOf(
-    R.string.tab_text_data,
-    R.string.tab_text_settings
-)
-
 /**
- * A [FragmentPagerAdapter] that returns a fragment corresponding to
+ * A [FragmentStateAdapter] that returns a fragment corresponding to
  * one of the sections/tabs/pages.
  */
-class SectionsPagerAdapter(private val context: Context, fm: FragmentManager) :
-    FragmentPagerAdapter(fm, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT) {
-
-    override fun getItem(position: Int): Fragment {
+class SectionsPagerAdapter(fa: FragmentActivity) : FragmentStateAdapter(fa) {
+    override fun createFragment(position: Int): Fragment {
         if (position == 0) {
-            return DataTabFragment.newInstance(position)
+            return DataTabFragment()
         }
         // getItem is called to instantiate the fragment for the given page.
         // Return a PlaceholderFragment (defined as a static inner class below).
-        return PlaceholderFragment.newInstance(position + 1)
+        return PlaceholderFragment()
     }
 
-    override fun getPageTitle(position: Int): CharSequence? {
-        return context.resources.getString(TAB_TITLES[position])
+    override fun getItemCount(): Int {
+        return TAB_TITLES.size
     }
 
-    override fun getCount(): Int {
-        // Show 2 total pages.
-        return 2
+    companion object {
+        private val TAB_TITLES = arrayOf(
+            R.string.tab_text_data,
+            R.string.tab_text_settings
+        )
+
+        fun getTabTitle(position: Int): Int {
+            return TAB_TITLES[position]
+        }
     }
 }
